Bugs fixed.

+++ train_main.py
@@ -3,14 +3,19 @@ import json
 import os
 import shutil
 
+import torch
 import yaml
 
 from models import create_model
 from utils import train_model, get_optimizer, get_loss, get_scheduler, test_model
 
 
-def resume(_config: dict):
+def train(_config, resume: bool = False):
     print(json.dumps(config, indent=4))
+
+    device = torch.device(_config['device'])
+    os.environ["CUDA_VISIBLE_DEVICES"] = str(device.index)
+
     dataset = _config['data']['dataset']
     model_name = _config['model']['name']
     optimizer_name = _config['optimizer']['name']
@@ -18,56 +23,13 @@ def resume(_config: dict):
 
     loss = get_loss(_config['loss']['name'])
 
-    loss.to(_config['train']['device'])
-
-    model, trainer = create_model(model_name,
-                                  dataset,
-                                  get_loss(_config['loss']['name']),
-                                  _config['model'][model_name],
-                                  _config['train']['device'])
-
-    optimizer = get_optimizer(optimizer_name, model.parameters(), **_config['optimizer'][optimizer_name])
-
-    scheduler = get_scheduler(scheduler_name, optimizer, **_config['scheduler'][scheduler_name])
-
-    save_folder = os.path.join('saves', dataset, _config['name'])
-
-    with open(os.path.join(save_folder, 'config.yaml'), 'w+') as _f:
-        yaml.safe_dump(_config, _f)
-
-    train_model(model=model,
-                dataset=dataset,
-                batch_size=_config['data']['batch-size'],
-                optimizer=optimizer,
-                scheduler=scheduler,
-                folder=save_folder,
-                trainer=trainer,
-                **_config['train'])
-
-    test_model(model=model,
-               dataset=dataset,
-               batch_size=_config['data']['batch-size'],
-               trainer=trainer,
-               folder=save_folder,
-               device=_config['train']['device'])
-
-
-def train(_config):
-    print(json.dumps(config, indent=4))
-    dataset = _config['data']['dataset']
-    model_name = _config['model']['name']
-    optimizer_name = _config['optimizer']['name']
-    scheduler_name = _config['scheduler']['name']
-
-    loss = get_loss(_config['loss']['name'])
-
-    loss.to(_config['train']['device'])
+    loss.to(device)
 
     model, trainer = create_model(model_name,
                                   dataset,
                                   loss,
                                   _config['model'][model_name],
-                                  _config['train']['device'])
+                                  device)
 
     optimizer = get_optimizer(optimizer_name, model.parameters(), **_config['optimizer'][optimizer_name])
 
@@ -75,8 +37,9 @@ def train(_config):
 
     save_folder = os.path.join('saves', dataset, _config['name'])
 
-    shutil.rmtree(save_folder, ignore_errors=True)
-    os.makedirs(save_folder)
+    if not resume:
+        shutil.rmtree(save_folder, ignore_errors=True)
+        os.makedirs(save_folder)
 
     with open(os.path.join(save_folder, 'config.yaml'), 'w+') as _f:
         yaml.safe_dump(_config, _f)
@@ -88,6 +51,7 @@ def train(_config):
                 scheduler=scheduler,
                 folder=save_folder,
                 trainer=trainer,
+                device=device,
                 **_config['train'])
 
     test_model(model=model,
@@ -95,7 +59,7 @@ def train(_config):
                batch_size=_config['data']['batch-size'],
                trainer=trainer,
                folder=save_folder,
-               device=_config['train']['device'])
+               device=device)
 
 
 if __name__ == '__main__':
@@ -111,6 +75,6 @@ if __name__ == '__main__':
         config = yaml.safe_load(f)
     if args.resume:
         print(f'Resume to {config["name"]}.')
-        resume(config)
+        train(config, resume=True)
     else:
         train(config)
